+++ tests/e2e/shell_command_outputs/mode_sync/__init__.py
@@ -5,7 +5,6 @@ from xmipp3_installer.installer.modes.mode_sync import mode_sync_executor
 
 from ... import shell_command_outputs
 
-TEST_FILES_DIR = os.path.join(".", "tests", "e2e", "test_files")
 SYNC_PROGRAM_NAME = f"{mode_sync_executor._SYNC_PROGRAM_NAME}.py"
 __NO_PROGRAM_INITIAL_MESSAGE = f"{os.path.join(mode_sync_executor._SYNC_PROGRAM_PATH, SYNC_PROGRAM_NAME)} does not exist."
 NO_PROGRAM = f"""{logger.red(__NO_PROGRAM_INITIAL_MESSAGE)}
