+++ tests/e2e/shell_command_outputs/mode_sync/mode_get_models.py
@@ -4,10 +4,11 @@ from xmipp3_installer.application.logger.logger import logger
 from xmipp3_installer.installer import constants, urls
 
 from .. import mode_sync
-from ...test_files import xmipp_sync_data
+from ....test_files import xmipp_sync_data
+from .... import TEST_FILES_DIR
 
 REAL_MODELS_DIR = os.path.join(constants.SOURCES_PATH, constants.XMIPP, "models")
-FAKE_SYNC_PROGRAM_FULL_PATH = os.path.join(mode_sync.TEST_FILES_DIR, mode_sync.SYNC_PROGRAM_NAME)
+FAKE_SYNC_PROGRAM_FULL_PATH = os.path.join(TEST_FILES_DIR, mode_sync.SYNC_PROGRAM_NAME)
 
 DOWNLOAD = '\n'.join([
 	"Downloading Deep Learning models (in background)",
