+++ tests/test_files/libraries-with-versions.txt
