+++ tests/test_files/xmipp_sync_data.py
